Update AsyncCommunicator.py
Modified averaging_standard() to rely on MPI.Recv instead of MPI.Irecv. 

Beforehand, averaging_standard() did not take into account that MIPI.Irecv may return False even when messages are available (but not yet transferred to the buffer). This may not allow the models to average themselves and it potentially generates segmentation faults.